kurang pdf dan dashboard

+++ resources/views/eskul/eskul.blade.php
@@ -22,14 +22,18 @@
     <div class="col-sm-12">
         <div class="card">
             <div class="card-body">
+            @if (auth()->user()->role_id == 1)
                 <a href="{{ url('eskul/add') }}" class="btn btn-primary mb-4"><i class="mdi mdi-plus"></i> Tambah</a>
+            @endif
                 <div class="table-responsive">
                     <table class="table user-table" id="table-datatables">
                         <thead>
                             <tr>
                                 <th>#</th>
                                 <th>Nama Eskul</th>
-                                <th>Action</th>
+                                @if (auth()->user()->role_id == 1)
+                                    <th>Action</th>
+                                @endif
                             </tr>
                         </thead>
                         <tbody>
@@ -37,11 +41,13 @@
                             <tr>
                                 <td>{{ $loop->iteration }}</td>
                                 <td> {{ $data->nama }}</td>
-                                <td>
-                                    <a href="eskulDetail/{{ $data->id }}" class="btn btn-success text-white"><i class="mdi mdi-plus"></i> Eskul Detail</a>
-                                    <a href="eskul/edit/{{ $data->id }}" class="btn btn-warning text-white"><i class="mdi mdi-pencil"></i> Edit</a>
-                                    <a href="eskul/hapus/{{ $data->id }}" class="btn btn-danger text-white" onclick="return confirm('Apakah anda yakin ingin menghapus data ini?')"><i class="mdi mdi-delete"></i> Hapus</a>
-                                </td>
+                                @if (auth()->user()->role_id == 1)
+                                    <td>
+                                        <a href="eskulDetail/{{ $data->id }}" class="btn btn-success text-white"><i class="mdi mdi-plus"></i> Eskul Detail</a>
+                                        <a href="eskul/edit/{{ $data->id }}" class="btn btn-warning text-white"><i class="mdi mdi-pencil"></i> Edit</a>
+                                        <a href="eskul/hapus/{{ $data->id }}" class="btn btn-danger text-white" onclick="return confirm('Apakah anda yakin ingin menghapus data ini?')"><i class="mdi mdi-delete"></i> Hapus</a>
+                                    </td>
+                                @endif
                             </tr>
                         @endforeach
                         </tbody>
